reorganize plotting code, add leafmap dep
